+++ src/main/java/yeogi/moim/security/UserDetailsServiceImpl.java
@@ -1,4 +1,4 @@
-package yeogi.moim.config;
+package yeogi.moim.security;
 
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
